raised state
