+++ src/App.js
@@ -5,11 +5,45 @@ import {
   Route,
   Link
 } from "react-router-dom";
+import axios from 'axios';
 import logo from './logo.svg';
 import './App.css';
 import CustomerList from './Components/CustomerList';
 
 class App extends Component {
+  constructor() {
+    super();
+
+    this.state = {
+      customers: [],
+      selectCustomer: undefined,
+      error: undefined,
+    }
+  }
+
+  componentDidMount() {
+    axios.get('http://localhost:3000/customers')
+      .then((response) => {
+        this.setState({
+          customers: response.data,
+        });
+      })
+      .catch((error) => {
+        this.setState({ error: error.message });
+      });
+  }
+
+  selectCustomer = ( customerId ) => {
+    console.log("selecting", customerId);
+    
+    const { customers } = this.state;
+    const selectedCustomer = customers.find((customer) => {
+      return customer.id === customerId;
+    });
+
+    this.setState({selectedCustomer});
+  }
+
   render() {
     return (
       <Router>
@@ -30,6 +64,8 @@ class App extends Component {
           </ul>
         </nav>
 
+        { this.state.error ? <div className="error-message">{this.state.error}</div> : "" }
+
         <Switch>
           <Route path="/search">
             <Search />
@@ -38,7 +74,7 @@ class App extends Component {
             <Library />
           </Route>
           <Route path="/customers">
-            <CustomerList />
+            <CustomerList selectCustomer={ this.selectCustomer } customers={ this.state.customers } />
           </Route>
           <Route path="/">
             <Home />
