+++ src/Components/CustomerList.js
@@ -1,81 +1,52 @@
-import React, { Component } from 'react';
-import axios from 'axios';
+import React from 'react';
 import PropTypes from 'prop-types';
 import CustomerRow from './CustomerRow';
 import './CustomerList.css';
 
-class CustomerList extends Component {
-  constructor(props) {
-    super(props);
+const CustomerList = (props) => {
 
-    this.state = {
-      customers: [],
-      error: undefined,
-      selectedCustomer: undefined,
-    };
+  const selectCustomer = ( customerId ) => {
+    this.props.selectCustomer(customerId);
   }
 
-  componentDidMount() {
-    axios.get('http://localhost:3000/customers')
-      .then((response) => {
-        this.setState({
-          customers: response.data,
-        });
-      })
-      .catch((error) => {
-        this.setState({ error: error.message });
-      });
-  }
-
-  selectCustomer = ( customerId ) => {
-    const { customers } = this.state;
-    const selectedCustomer = customers.find((customer) => {
-      return customer.id === customerId;
-    });
-
-    this.setState({selectedCustomer});
-  }
-
-  makeCustomerList = (customers) => {
+  const makeCustomerList = (customers) => {
     return customers.map((customer, i) => {
       return <CustomerRow 
         {...customer}
         key={i}
-        selectCustomer={ this.selectCustomer }
+        selectCustomer={ props.selectCustomer }
       />
     });
   }
 
-  render() {
-    return (
-      <section>
-        <h1>Customers</h1>
-        { this.state.error ? <div className="error-message">{this.state.error}</div> : "" }
-        <table className="customer-list">
-          <thead>
-            <tr>
-              <th>ID</th>
-              <th>Name</th>
-              <th>Movies Checked Out</th>
-              <th>Account Credit</th>
-              <th>Address</th>
-              <th>Phone Number</th>
-              <th>Registrated Date</th>
-              <th>Select Customer</th>
-            </tr>
-          </thead>
-          <tbody>
-            { this.makeCustomerList(this.state.customers) }
-          </tbody>
-        </table>
+  return (
+    <section>
+      <h1>Customers</h1>
+      <table className="customer-list">
+        <thead>
+          <tr>
+            <th>ID</th>
+            <th>Name</th>
+            <th>Movies Checked Out</th>
+            <th>Account Credit</th>
+            <th>Address</th>
+            <th>Phone Number</th>
+            <th>Registrated Date</th>
+            <th>Select Customer</th>
+          </tr>
+        </thead>
+        <tbody>
+          { makeCustomerList(props.customers) }
+        </tbody>
+      </table>
 
-      </section>
-    );
-  }
+    </section>
+  );
 }
 
 CustomerList.propTypes = {
-
+  selectCustomer: PropTypes.func.isRequired,
+  customers: PropTypes.array.isRequired,
 }
 
 export default CustomerList;
